fix: 도메인 수정

+++ src/main/java/com/example/airbnb_clonecoding/model/Room.java
@@ -13,7 +13,7 @@ public class Room {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ROOM_ID")
-    private Long roomId;
+    private int roomId;
 
     @ManyToOne
     private Host hostId;
